requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
deepspeed==0.16.4
einops==0.8.1
faiss-cpu==1.10.0
filelock==3.13.1
flash_attn==2.7.4.post1
fsspec==2024.6.1
hjson==3.1.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
ninja==1.11.1.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pillow==11.0.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==75.8.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
wheel==0.45.1
